// frontend/src/app/model/user.model.ts
// user.model.ts
export interface User {
    id: number;
    first_name: string;
    middle_name: string;
    last_name: string;
    position: string;
    email: string;
    password: string;
    address_line_1: string;
    city: string;
    barangay: string;
    contact_number: string;
    image: string;
    role: string;
}
  